Reverted from separatated tpdata module
It seems like having all the data in a separated module
breaks the whole thing for some obsure autoloading issues.

The approach to a separated tpdata module will be retried
in the future, for the moment we keep on with a
all in one setup to avoid developments delays.

Relevant errors with a separated tpdata module:

During spec tests:
     Failure/Error: })
     Puppet::Error:
       Could not find module tpdata in environment production at /Users/al/Documents/github/EXAMPLE42/puppet-modules/tp/spec/fixtures/modules/tp/manifests/conf.pp:36 on node macante.local
     # ./spec/fixtures/modules/tp/lib/puppet/parser/functions/tp_lookup.rb:55:in `block in <module:Functions>'
     # ./spec/defines/tp_conf_spec.rb:250:in `block (3 levels) in <top (required)>'

(yes, tpdata was in fixtures)

With puppet apply:
Error: Could not find module tpdata in environment production on node
macante.local
/Users/al/Documents/github/EXAMPLE42/puppet-modules/tp/lib/puppet/parser/functions/tp_lookup.rb:55:in
`block in <module:Functions>'
/Users/al/.rvm/gems/ruby-1.9.3-p392@puppet_spec/gems/puppet-3.6.1/lib/puppet/parser/functions.rb:155:in
`block (2 levels) in newfunction'
/Users/al/.rvm/gems/ruby-1.9.3-p392@puppet_spec/gems/puppet-3.6.1/lib/puppet/util/profiler/none.rb:6:in
`profile'

Per aspera ad astra